Fix ANY queries and TCP socket for DNS
Sets up both TCP and UDP DNS servers, which should allow for the
occasional case where a TCP query can make it to the server but UDP
can't, and fixes support for ANY queries. ANY queries are deprecated in
RFC8482 but this may also provide for additional interaction
opportunities.

Re-organises the error handling around DNS server creation slightly
because 'error' was always returned as nil anyway.