Implement IPv6 support for DNS, fix some DNS server behaviours
This commit allows the DNS server to serve IPv6 address records (AAAA)
for both interaction and NS glue. In combination with the changes in
PRs #181 and #182 this allows for IPv6 interaction and provides more
opportunities for interaction to ocurr when e.g. IPv6 and IPv4 egress
firewalling differ on a network being tested.

Specifying the '-ip' option multiple times now allows setting multiple
response addresses for the DNS server. Both IPv4 and IPv6 addresses are
supported. If no IP address is specified the auto-detected address will
be logged at startup time to aid with troubleshooting.

Fixes a couple of behaviour bugs with the DNS server implementation and
changes the default TTL to 5 from 3600.